Nora: Fix DR mode is changed when not OTG
Description:
The USB DR mode should not be changed by data role that detected
by PD if the DR mode is forced to host or peripheral.

Test:
Change the dr_mode in DTS to "host". No ADB device is shown
when connected to PC. But connect USB disk to Nora can be detected.

Change-Id: I93b6225600d2710aba047f8d1bb22042681a6348